requirements.txt
flask==2.3.3
flask-cors==4.0.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
requests==2.31.0
python-dotenv==1.0.0
boto3==1.28.40
beautifulsoup4==4.12.2
oauthlib
psutil
requests-oauthlib
openai
psycopg2-binary
